Polishing some code

+++ src/main/java/org/apache/ibatis/type/ByteArrayUtils.java
@@ -27,7 +27,7 @@ class ByteArrayUtils {
   static byte[] convertToPrimitiveArray(Byte[] objects) {
     final byte[] bytes = new byte[objects.length];
     for (int i = 0; i < objects.length; i++) {
-      bytes[i] = objects[i].byteValue();
+      bytes[i] = objects[i];
     }
     return bytes;
   }
@@ -35,7 +35,7 @@ class ByteArrayUtils {
   static Byte[] convertToObjectArray(byte[] bytes) {
     final Byte[] objects = new Byte[bytes.length];
     for (int i = 0; i < bytes.length; i++) {
-      objects[i] = Byte.valueOf(bytes[i]);
+      objects[i] = bytes[i];
     }
     return objects;
   }
